Divide the test for exposed path into more small tests
This avoids a giant conflict for future change.
