+++ packages/api_tests/__tests__/exposed_path/ClassicOption.test.mjs
@@ -1,0 +1,11 @@
+import test from 'ava';
+import { testImportExposedPath } from './helper.mjs';
+
+testImportExposedPath(
+    test,
+    Object.keys({
+        './cjs/ClassicOption': null,
+        './esm/ClassicOption': null,
+        './lib/ClassicOption': null,
+    })
+);
