+++ packages/api_tests/__tests__/exposed_path/Maybe.test.mjs
@@ -1,0 +1,126 @@
+import test from 'ava';
+import { testImportExposedPath } from './helper.mjs';
+
+testImportExposedPath(
+    test,
+    Object.keys({
+        './cjs/Maybe/ErrorMessage': null,
+        './cjs/Maybe/Maybe': null,
+        './cjs/Maybe/and': null,
+        './cjs/Maybe/andThen': null,
+        './cjs/Maybe/andThenAsync': null,
+        './cjs/Maybe/compat/v33': null,
+        './cjs/Maybe/expect': null,
+        './cjs/Maybe/inspect': null,
+        './cjs/Maybe': null,
+        './cjs/Maybe/map': null,
+        './cjs/Maybe/mapAsync': null,
+        './cjs/Maybe/mapOr': null,
+        './cjs/Maybe/mapOrAsync': null,
+        './cjs/Maybe/mapOrElse': null,
+        './cjs/Maybe/mapOrElseAsync': null,
+        './cjs/Maybe/namespace': null,
+        './cjs/Maybe/okOr': null,
+        './cjs/Maybe/okOrElse': null,
+        './cjs/Maybe/or': null,
+        './cjs/Maybe/orElse': null,
+        './cjs/Maybe/orElseAsync': null,
+        './cjs/Maybe/toNullable': null,
+        './cjs/Maybe/toPlainResult': null,
+        './cjs/Maybe/toUndefinable': null,
+        './cjs/Maybe/unwrap': null,
+        './cjs/Maybe/unwrapOr': null,
+        './cjs/Maybe/unwrapOrElse': null,
+        './cjs/Maybe/unwrapOrElseAsync': null,
+        './cjs/Maybe/xor': null,
+
+        './esm/Maybe/ErrorMessage': null,
+        './esm/Maybe/Maybe': null,
+        './esm/Maybe/and': null,
+        './esm/Maybe/andThen': null,
+        './esm/Maybe/andThenAsync': null,
+        './esm/Maybe/compat/v33': null,
+        './esm/Maybe/expect': null,
+        './esm/Maybe/inspect': null,
+        './esm/Maybe': null,
+        './esm/Maybe/map': null,
+        './esm/Maybe/mapAsync': null,
+        './esm/Maybe/mapOr': null,
+        './esm/Maybe/mapOrAsync': null,
+        './esm/Maybe/mapOrElse': null,
+        './esm/Maybe/mapOrElseAsync': null,
+        './esm/Maybe/namespace': null,
+        './esm/Maybe/okOr': null,
+        './esm/Maybe/okOrElse': null,
+        './esm/Maybe/or': null,
+        './esm/Maybe/orElse': null,
+        './esm/Maybe/orElseAsync': null,
+        './esm/Maybe/toNullable': null,
+        './esm/Maybe/toPlainResult': null,
+        './esm/Maybe/toUndefinable': null,
+        './esm/Maybe/unwrap': null,
+        './esm/Maybe/unwrapOr': null,
+        './esm/Maybe/unwrapOrElse': null,
+        './esm/Maybe/unwrapOrElseAsync': null,
+        './esm/Maybe/xor': null,
+
+        './lib/Maybe/ErrorMessage': null,
+        './lib/Maybe/Maybe': null,
+        './lib/Maybe/and': null,
+        './lib/Maybe/andThen': null,
+        './lib/Maybe/andThenAsync': null,
+        './lib/Maybe/compat/v33': null,
+        './lib/Maybe/expect': null,
+        './lib/Maybe/inspect': null,
+        './lib/Maybe': null,
+        './lib/Maybe/map': null,
+        './lib/Maybe/mapAsync': null,
+        './lib/Maybe/mapOr': null,
+        './lib/Maybe/mapOrAsync': null,
+        './lib/Maybe/mapOrElse': null,
+        './lib/Maybe/mapOrElseAsync': null,
+        './lib/Maybe/namespace': null,
+        './lib/Maybe/okOr': null,
+        './lib/Maybe/okOrElse': null,
+        './lib/Maybe/or': null,
+        './lib/Maybe/orElse': null,
+        './lib/Maybe/orElseAsync': null,
+        './lib/Maybe/toNullable': null,
+        './lib/Maybe/toPlainResult': null,
+        './lib/Maybe/toUndefinable': null,
+        './lib/Maybe/unwrap': null,
+        './lib/Maybe/unwrapOr': null,
+        './lib/Maybe/unwrapOrElse': null,
+        './lib/Maybe/unwrapOrElseAsync': null,
+        './lib/Maybe/xor': null,
+
+        './Maybe': null,
+        './Maybe/and': null,
+        './Maybe/andThen': null,
+        './Maybe/andThenAsync': null,
+        './Maybe/compat/v33': null,
+        './Maybe/expect': null,
+        './Maybe/inspect': null,
+        './Maybe/map': null,
+        './Maybe/mapAsync': null,
+        './Maybe/Maybe': null,
+        './Maybe/mapOr': null,
+        './Maybe/mapOrAsync': null,
+        './Maybe/mapOrElse': null,
+        './Maybe/mapOrElseAsync': null,
+        './Maybe/namespace': null,
+        './Maybe/okOr': null,
+        './Maybe/okOrElse': null,
+        './Maybe/or': null,
+        './Maybe/orElse': null,
+        './Maybe/orElseAsync': null,
+        './Maybe/toNullable': null,
+        './Maybe/toPlainResult': null,
+        './Maybe/toUndefinable': null,
+        './Maybe/unwrap': null,
+        './Maybe/unwrapOr': null,
+        './Maybe/unwrapOrElse': null,
+        './Maybe/unwrapOrElseAsync': null,
+        './Maybe/xor': null,
+    })
+);
