+++ packages/api_tests/__tests__/exposed_path/PlainOption.test.mjs
@@ -1,0 +1,143 @@
+import test from 'ava';
+import { testImportExposedPath } from './helper.mjs';
+
+testImportExposedPath(
+    test,
+    Object.keys({
+        './cjs/PlainOption/Option': null,
+        './cjs/PlainOption/and': null,
+        './cjs/PlainOption/andThen': null,
+        './cjs/PlainOption/andThenAsync': null,
+        './cjs/PlainOption/asMut': null,
+        './cjs/PlainOption/compat/v33': null,
+        './cjs/PlainOption/drop': null,
+        './cjs/PlainOption/equal': null,
+        './cjs/PlainOption/expect': null,
+        './cjs/PlainOption/filter': null,
+        './cjs/PlainOption/flatten': null,
+        './cjs/PlainOption/inspect': null,
+        './cjs/PlainOption': null,
+        './cjs/PlainOption/map': null,
+        './cjs/PlainOption/mapAsync': null,
+        './cjs/PlainOption/mapOr': null,
+        './cjs/PlainOption/mapOrAsync': null,
+        './cjs/PlainOption/mapOrElse': null,
+        './cjs/PlainOption/mapOrElseAsync': null,
+        './cjs/PlainOption/namespace': null,
+        './cjs/PlainOption/okOr': null,
+        './cjs/PlainOption/okOrElse': null,
+        './cjs/PlainOption/or': null,
+        './cjs/PlainOption/orElse': null,
+        './cjs/PlainOption/orElseAsync': null,
+        './cjs/PlainOption/transpose': null,
+        './cjs/PlainOption/toNullable': null,
+        './cjs/PlainOption/toUndefinable': null,
+        './cjs/PlainOption/unwrap': null,
+        './cjs/PlainOption/unwrapOr': null,
+        './cjs/PlainOption/unwrapOrElse': null,
+        './cjs/PlainOption/unwrapOrElseAsync': null,
+        './cjs/PlainOption/xor': null,
+
+        './esm/PlainOption/Option': null,
+        './esm/PlainOption/and': null,
+        './esm/PlainOption/andThen': null,
+        './esm/PlainOption/andThenAsync': null,
+        './esm/PlainOption/asMut': null,
+        './esm/PlainOption/compat/v33': null,
+        './esm/PlainOption/drop': null,
+        './esm/PlainOption/equal': null,
+        './esm/PlainOption/expect': null,
+        './esm/PlainOption/filter': null,
+        './esm/PlainOption/flatten': null,
+        './esm/PlainOption/inspect': null,
+        './esm/PlainOption': null,
+        './esm/PlainOption/map': null,
+        './esm/PlainOption/mapAsync': null,
+        './esm/PlainOption/mapOr': null,
+        './esm/PlainOption/mapOrAsync': null,
+        './esm/PlainOption/mapOrElse': null,
+        './esm/PlainOption/mapOrElseAsync': null,
+        './esm/PlainOption/namespace': null,
+        './esm/PlainOption/okOr': null,
+        './esm/PlainOption/okOrElse': null,
+        './esm/PlainOption/or': null,
+        './esm/PlainOption/orElse': null,
+        './esm/PlainOption/orElseAsync': null,
+        './esm/PlainOption/transpose': null,
+        './esm/PlainOption/toNullable': null,
+        './esm/PlainOption/toUndefinable': null,
+        './esm/PlainOption/unwrap': null,
+        './esm/PlainOption/unwrapOr': null,
+        './esm/PlainOption/unwrapOrElse': null,
+        './esm/PlainOption/unwrapOrElseAsync': null,
+        './esm/PlainOption/xor': null,
+
+        './lib/PlainOption/Option': null,
+        './lib/PlainOption/and': null,
+        './lib/PlainOption/andThen': null,
+        './lib/PlainOption/andThenAsync': null,
+        './lib/PlainOption/asMut': null,
+        './lib/PlainOption/compat/v33': null,
+        './lib/PlainOption/drop': null,
+        './lib/PlainOption/expect': null,
+        './lib/PlainOption/equal': null,
+        './lib/PlainOption/flatten': null,
+        './lib/PlainOption/filter': null,
+        './lib/PlainOption/inspect': null,
+        './lib/PlainOption': null,
+        './lib/PlainOption/map': null,
+        './lib/PlainOption/mapAsync': null,
+        './lib/PlainOption/mapOr': null,
+        './lib/PlainOption/mapOrAsync': null,
+        './lib/PlainOption/mapOrElse': null,
+        './lib/PlainOption/mapOrElseAsync': null,
+        './lib/PlainOption/namespace': null,
+        './lib/PlainOption/okOr': null,
+        './lib/PlainOption/okOrElse': null,
+        './lib/PlainOption/or': null,
+        './lib/PlainOption/orElse': null,
+        './lib/PlainOption/orElseAsync': null,
+        './lib/PlainOption/transpose': null,
+        './lib/PlainOption/toNullable': null,
+        './lib/PlainOption/toUndefinable': null,
+        './lib/PlainOption/unwrap': null,
+        './lib/PlainOption/unwrapOr': null,
+        './lib/PlainOption/unwrapOrElse': null,
+        './lib/PlainOption/unwrapOrElseAsync': null,
+        './lib/PlainOption/xor': null,
+
+        './PlainOption': null,
+        './PlainOption/and': null,
+        './PlainOption/andThen': null,
+        './PlainOption/andThenAsync': null,
+        './PlainOption/asMut': null,
+        './PlainOption/compat/v33': null,
+        './PlainOption/drop': null,
+        './PlainOption/equal': null,
+        './PlainOption/expect': null,
+        './PlainOption/filter': null,
+        './PlainOption/flatten': null,
+        './PlainOption/inspect': null,
+        './PlainOption/map': null,
+        './PlainOption/mapAsync': null,
+        './PlainOption/mapOr': null,
+        './PlainOption/mapOrAsync': null,
+        './PlainOption/mapOrElse': null,
+        './PlainOption/mapOrElseAsync': null,
+        './PlainOption/okOr': null,
+        './PlainOption/okOrElse': null,
+        './PlainOption/Option': null,
+        './PlainOption/namespace': null,
+        './PlainOption/or': null,
+        './PlainOption/orElse': null,
+        './PlainOption/orElseAsync': null,
+        './PlainOption/transpose': null,
+        './PlainOption/toNullable': null,
+        './PlainOption/toUndefinable': null,
+        './PlainOption/unwrap': null,
+        './PlainOption/unwrapOr': null,
+        './PlainOption/unwrapOrElse': null,
+        './PlainOption/unwrapOrElseAsync': null,
+        './PlainOption/xor': null,
+    })
+);
