+++ packages/api_tests/__tests__/exposed_path/PlainResult.test.mjs
@@ -1,0 +1,155 @@
+import test from 'ava';
+import { testImportExposedPath } from './helper.mjs';
+
+testImportExposedPath(
+    test,
+    Object.keys({
+        './cjs/PlainResult/Result': null,
+        './cjs/PlainResult/and': null,
+        './cjs/PlainResult/andThen': null,
+        './cjs/PlainResult/andThenAsync': null,
+        './cjs/PlainResult/asMut': null,
+        './cjs/PlainResult/compat/v33': null,
+        './cjs/PlainResult/drop': null,
+        './cjs/PlainResult/equal': null,
+        './cjs/PlainResult/expect': null,
+        './cjs/PlainResult/flatten': null,
+        './cjs/PlainResult/fromPromiseSettledResult': null,
+        './cjs/PlainResult/inspect': null,
+        './cjs/PlainResult': null,
+        './cjs/PlainResult/map': null,
+        './cjs/PlainResult/mapAsync': null,
+        './cjs/PlainResult/mapErr': null,
+        './cjs/PlainResult/mapErrAsync': null,
+        './cjs/PlainResult/mapOr': null,
+        './cjs/PlainResult/mapOrAsync': null,
+        './cjs/PlainResult/mapOrElse': null,
+        './cjs/PlainResult/mapOrElseAsync': null,
+        './cjs/PlainResult/namespace': null,
+        './cjs/PlainResult/or': null,
+        './cjs/PlainResult/orElse': null,
+        './cjs/PlainResult/orElseAsync': null,
+        './cjs/PlainResult/transpose': null,
+        './cjs/PlainResult/toNullable': null,
+        './cjs/PlainResult/toOption': null,
+        './cjs/PlainResult/toUndefinable': null,
+        './cjs/PlainResult/tryCatch': null,
+        './cjs/PlainResult/tryCatchAsync': null,
+        './cjs/PlainResult/unwrap': null,
+        './cjs/PlainResult/unwrapOr': null,
+        './cjs/PlainResult/unwrapOrElse': null,
+        './cjs/PlainResult/unwrapOrElseAsync': null,
+        './cjs/PlainResult/unwrapOrThrowError': null,
+
+        './esm/PlainResult/Result': null,
+        './esm/PlainResult/and': null,
+        './esm/PlainResult/andThen': null,
+        './esm/PlainResult/andThenAsync': null,
+        './esm/PlainResult/asMut': null,
+        './esm/PlainResult/compat/v33': null,
+        './esm/PlainResult/drop': null,
+        './esm/PlainResult/equal': null,
+        './esm/PlainResult/expect': null,
+        './esm/PlainResult/flatten': null,
+        './esm/PlainResult/fromPromiseSettledResult': null,
+        './esm/PlainResult/inspect': null,
+        './esm/PlainResult': null,
+        './esm/PlainResult/map': null,
+        './esm/PlainResult/mapAsync': null,
+        './esm/PlainResult/mapErr': null,
+        './esm/PlainResult/mapErrAsync': null,
+        './esm/PlainResult/mapOr': null,
+        './esm/PlainResult/mapOrAsync': null,
+        './esm/PlainResult/mapOrElse': null,
+        './esm/PlainResult/mapOrElseAsync': null,
+        './esm/PlainResult/namespace': null,
+        './esm/PlainResult/or': null,
+        './esm/PlainResult/orElse': null,
+        './esm/PlainResult/orElseAsync': null,
+        './esm/PlainResult/transpose': null,
+        './esm/PlainResult/toNullable': null,
+        './esm/PlainResult/toOption': null,
+        './esm/PlainResult/toUndefinable': null,
+        './esm/PlainResult/tryCatch': null,
+        './esm/PlainResult/tryCatchAsync': null,
+        './esm/PlainResult/unwrap': null,
+        './esm/PlainResult/unwrapOr': null,
+        './esm/PlainResult/unwrapOrElse': null,
+        './esm/PlainResult/unwrapOrElseAsync': null,
+        './esm/PlainResult/unwrapOrThrowError': null,
+
+        './lib/PlainResult/Result': null,
+        './lib/PlainResult/and': null,
+        './lib/PlainResult/andThen': null,
+        './lib/PlainResult/andThenAsync': null,
+        './lib/PlainResult/asMut': null,
+        './lib/PlainResult/compat/v33': null,
+        './lib/PlainResult/drop': null,
+        './lib/PlainResult/equal': null,
+        './lib/PlainResult/expect': null,
+        './lib/PlainResult/flatten': null,
+        './lib/PlainResult/fromPromiseSettledResult': null,
+        './lib/PlainResult/inspect': null,
+        './lib/PlainResult': null,
+        './lib/PlainResult/map': null,
+        './lib/PlainResult/mapAsync': null,
+        './lib/PlainResult/mapErr': null,
+        './lib/PlainResult/mapErrAsync': null,
+        './lib/PlainResult/mapOr': null,
+        './lib/PlainResult/mapOrAsync': null,
+        './lib/PlainResult/mapOrElse': null,
+        './lib/PlainResult/mapOrElseAsync': null,
+        './lib/PlainResult/namespace': null,
+        './lib/PlainResult/or': null,
+        './lib/PlainResult/orElse': null,
+        './lib/PlainResult/orElseAsync': null,
+        './lib/PlainResult/transpose': null,
+        './lib/PlainResult/toNullable': null,
+        './lib/PlainResult/toOption': null,
+        './lib/PlainResult/toUndefinable': null,
+        './lib/PlainResult/tryCatch': null,
+        './lib/PlainResult/tryCatchAsync': null,
+        './lib/PlainResult/unwrap': null,
+        './lib/PlainResult/unwrapOr': null,
+        './lib/PlainResult/unwrapOrElse': null,
+        './lib/PlainResult/unwrapOrElseAsync': null,
+        './lib/PlainResult/unwrapOrThrowError': null,
+
+        './PlainResult': null,
+        './PlainResult/and': null,
+        './PlainResult/andThen': null,
+        './PlainResult/andThenAsync': null,
+        './PlainResult/asMut': null,
+        './PlainResult/compat/v33': null,
+        './PlainResult/drop': null,
+        './PlainResult/equal': null,
+        './PlainResult/expect': null,
+        './PlainResult/flatten': null,
+        './PlainResult/fromPromiseSettledResult': null,
+        './PlainResult/inspect': null,
+        './PlainResult/map': null,
+        './PlainResult/mapAsync': null,
+        './PlainResult/mapErr': null,
+        './PlainResult/mapErrAsync': null,
+        './PlainResult/mapOr': null,
+        './PlainResult/mapOrAsync': null,
+        './PlainResult/mapOrElse': null,
+        './PlainResult/mapOrElseAsync': null,
+        './PlainResult/namespace': null,
+        './PlainResult/or': null,
+        './PlainResult/orElse': null,
+        './PlainResult/orElseAsync': null,
+        './PlainResult/Result': null,
+        './PlainResult/toNullable': null,
+        './PlainResult/toOption': null,
+        './PlainResult/toUndefinable': null,
+        './PlainResult/tryCatch': null,
+        './PlainResult/tryCatchAsync': null,
+        './PlainResult/transpose': null,
+        './PlainResult/unwrap': null,
+        './PlainResult/unwrapOr': null,
+        './PlainResult/unwrapOrElse': null,
+        './PlainResult/unwrapOrElseAsync': null,
+        './PlainResult/unwrapOrThrowError': null,
+    })
+);
