+++ packages/api_tests/__tests__/exposed_path/index.test.mjs
@@ -1,0 +1,9 @@
+import test from 'ava';
+import { testImportExposedPath } from './helper.mjs';
+
+testImportExposedPath(
+    test,
+    Object.keys({
+        '.': null,
+    })
+);
