--- packages/api_tests/__tests__/exposed_path_table.mjs
@@ -1,622 +1,0 @@
-export const EXPOSED_PATH_LIST = Object.keys({
-    './cjs/ClassicOption': null,
-    './cjs/ClassicResult': null,
-    './cjs/Maybe/ErrorMessage': null,
-    './cjs/Maybe/Maybe': null,
-    './cjs/Maybe/and': null,
-    './cjs/Maybe/andThen': null,
-    './cjs/Maybe/andThenAsync': null,
-    './cjs/Maybe/compat/v33': null,
-    './cjs/Maybe/expect': null,
-    './cjs/Maybe/inspect': null,
-    './cjs/Maybe': null,
-    './cjs/Maybe/map': null,
-    './cjs/Maybe/mapAsync': null,
-    './cjs/Maybe/mapOr': null,
-    './cjs/Maybe/mapOrAsync': null,
-    './cjs/Maybe/mapOrElse': null,
-    './cjs/Maybe/mapOrElseAsync': null,
-    './cjs/Maybe/namespace': null,
-    './cjs/Maybe/okOr': null,
-    './cjs/Maybe/okOrElse': null,
-    './cjs/Maybe/or': null,
-    './cjs/Maybe/orElse': null,
-    './cjs/Maybe/orElseAsync': null,
-    './cjs/Maybe/toNullable': null,
-    './cjs/Maybe/toPlainResult': null,
-    './cjs/Maybe/toUndefinable': null,
-    './cjs/Maybe/unwrap': null,
-    './cjs/Maybe/unwrapOr': null,
-    './cjs/Maybe/unwrapOrElse': null,
-    './cjs/Maybe/unwrapOrElseAsync': null,
-    './cjs/Maybe/xor': null,
-    './cjs/Nullable/ErrorMessage': null,
-    './cjs/Nullable/Nullable': null,
-    './cjs/Nullable/and': null,
-    './cjs/Nullable/andThen': null,
-    './cjs/Nullable/andThenAsync': null,
-    './cjs/Nullable/compat/v33': null,
-    './cjs/Nullable/expect': null,
-    './cjs/Nullable/inspect': null,
-    './cjs/Nullable': null,
-    './cjs/Nullable/map': null,
-    './cjs/Nullable/mapAsync': null,
-    './cjs/Nullable/mapOr': null,
-    './cjs/Nullable/mapOrAsync': null,
-    './cjs/Nullable/mapOrElse': null,
-    './cjs/Nullable/mapOrElseAsync': null,
-    './cjs/Nullable/namespace': null,
-    './cjs/Nullable/okOr': null,
-    './cjs/Nullable/okOrElse': null,
-    './cjs/Nullable/or': null,
-    './cjs/Nullable/orElse': null,
-    './cjs/Nullable/orElseAsync': null,
-    './cjs/Nullable/toPlainResult': null,
-    './cjs/Nullable/toUndefinable': null,
-    './cjs/Nullable/unwrap': null,
-    './cjs/Nullable/unwrapOr': null,
-    './cjs/Nullable/unwrapOrElse': null,
-    './cjs/Nullable/unwrapOrElseAsync': null,
-    './cjs/Nullable/xor': null,
-    './cjs/PlainOption/Option': null,
-    './cjs/PlainOption/and': null,
-    './cjs/PlainOption/andThen': null,
-    './cjs/PlainOption/andThenAsync': null,
-    './cjs/PlainOption/asMut': null,
-    './cjs/PlainOption/compat/v33': null,
-    './cjs/PlainOption/drop': null,
-    './cjs/PlainOption/equal': null,
-    './cjs/PlainOption/expect': null,
-    './cjs/PlainOption/filter': null,
-    './cjs/PlainOption/flatten': null,
-    './cjs/PlainOption/inspect': null,
-    './cjs/PlainOption': null,
-    './cjs/PlainOption/map': null,
-    './cjs/PlainOption/mapAsync': null,
-    './cjs/PlainOption/mapOr': null,
-    './cjs/PlainOption/mapOrAsync': null,
-    './cjs/PlainOption/mapOrElse': null,
-    './cjs/PlainOption/mapOrElseAsync': null,
-    './cjs/PlainOption/namespace': null,
-    './cjs/PlainOption/okOr': null,
-    './cjs/PlainOption/okOrElse': null,
-    './cjs/PlainOption/or': null,
-    './cjs/PlainOption/orElse': null,
-    './cjs/PlainOption/orElseAsync': null,
-    './cjs/PlainOption/transpose': null,
-    './cjs/PlainOption/toNullable': null,
-    './cjs/PlainOption/toUndefinable': null,
-    './cjs/PlainOption/unwrap': null,
-    './cjs/PlainOption/unwrapOr': null,
-    './cjs/PlainOption/unwrapOrElse': null,
-    './cjs/PlainOption/unwrapOrElseAsync': null,
-    './cjs/PlainOption/xor': null,
-    './cjs/PlainResult/Result': null,
-    './cjs/PlainResult/and': null,
-    './cjs/PlainResult/andThen': null,
-    './cjs/PlainResult/andThenAsync': null,
-    './cjs/PlainResult/asMut': null,
-    './cjs/PlainResult/compat/v33': null,
-    './cjs/PlainResult/drop': null,
-    './cjs/PlainResult/equal': null,
-    './cjs/PlainResult/expect': null,
-    './cjs/PlainResult/flatten': null,
-    './cjs/PlainResult/fromPromiseSettledResult': null,
-    './cjs/PlainResult/inspect': null,
-    './cjs/PlainResult': null,
-    './cjs/PlainResult/map': null,
-    './cjs/PlainResult/mapAsync': null,
-    './cjs/PlainResult/mapErr': null,
-    './cjs/PlainResult/mapErrAsync': null,
-    './cjs/PlainResult/mapOr': null,
-    './cjs/PlainResult/mapOrAsync': null,
-    './cjs/PlainResult/mapOrElse': null,
-    './cjs/PlainResult/mapOrElseAsync': null,
-    './cjs/PlainResult/namespace': null,
-    './cjs/PlainResult/or': null,
-    './cjs/PlainResult/orElse': null,
-    './cjs/PlainResult/orElseAsync': null,
-    './cjs/PlainResult/transpose': null,
-    './cjs/PlainResult/toNullable': null,
-    './cjs/PlainResult/toOption': null,
-    './cjs/PlainResult/toUndefinable': null,
-    './cjs/PlainResult/tryCatch': null,
-    './cjs/PlainResult/tryCatchAsync': null,
-    './cjs/PlainResult/unwrap': null,
-    './cjs/PlainResult/unwrapOr': null,
-    './cjs/PlainResult/unwrapOrElse': null,
-    './cjs/PlainResult/unwrapOrElseAsync': null,
-    './cjs/PlainResult/unwrapOrThrowError': null,
-    './cjs/Undefinable/ErrorMessage': null,
-    './cjs/Undefinable/Undefinable': null,
-    './cjs/Undefinable/and': null,
-    './cjs/Undefinable/andThen': null,
-    './cjs/Undefinable/andThenAsync': null,
-    './cjs/Undefinable/expect': null,
-    './cjs/Undefinable/compat/v33': null,
-    './cjs/Undefinable/inspect': null,
-    './cjs/Undefinable': null,
-    './cjs/Undefinable/map': null,
-    './cjs/Undefinable/mapAsync': null,
-    './cjs/Undefinable/mapOr': null,
-    './cjs/Undefinable/mapOrAsync': null,
-    './cjs/Undefinable/mapOrElse': null,
-    './cjs/Undefinable/mapOrElseAsync': null,
-    './cjs/Undefinable/namespace': null,
-    './cjs/Undefinable/okOr': null,
-    './cjs/Undefinable/okOrElse': null,
-    './cjs/Undefinable/or': null,
-    './cjs/Undefinable/orElse': null,
-    './cjs/Undefinable/orElseAsync': null,
-    './cjs/Undefinable/toNullable': null,
-    './cjs/Undefinable/toPlainResult': null,
-    './cjs/Undefinable/unwrap': null,
-    './cjs/Undefinable/unwrapOr': null,
-    './cjs/Undefinable/unwrapOrElse': null,
-    './cjs/Undefinable/unwrapOrElseAsync': null,
-    './cjs/Undefinable/xor': null,
-    './esm/ClassicOption': null,
-    './esm/ClassicResult': null,
-    './esm/Maybe/ErrorMessage': null,
-    './esm/Maybe/Maybe': null,
-    './esm/Maybe/and': null,
-    './esm/Maybe/andThen': null,
-    './esm/Maybe/andThenAsync': null,
-    './esm/Maybe/compat/v33': null,
-    './esm/Maybe/expect': null,
-    './esm/Maybe/inspect': null,
-    './esm/Maybe': null,
-    './esm/Maybe/map': null,
-    './esm/Maybe/mapAsync': null,
-    './esm/Maybe/mapOr': null,
-    './esm/Maybe/mapOrAsync': null,
-    './esm/Maybe/mapOrElse': null,
-    './esm/Maybe/mapOrElseAsync': null,
-    './esm/Maybe/namespace': null,
-    './esm/Maybe/okOr': null,
-    './esm/Maybe/okOrElse': null,
-    './esm/Maybe/or': null,
-    './esm/Maybe/orElse': null,
-    './esm/Maybe/orElseAsync': null,
-    './esm/Maybe/toNullable': null,
-    './esm/Maybe/toPlainResult': null,
-    './esm/Maybe/toUndefinable': null,
-    './esm/Maybe/unwrap': null,
-    './esm/Maybe/unwrapOr': null,
-    './esm/Maybe/unwrapOrElse': null,
-    './esm/Maybe/unwrapOrElseAsync': null,
-    './esm/Maybe/xor': null,
-    './esm/Nullable/ErrorMessage': null,
-    './esm/Nullable/Nullable': null,
-    './esm/Nullable/and': null,
-    './esm/Nullable/andThen': null,
-    './esm/Nullable/andThenAsync': null,
-    './esm/Nullable/compat/v33': null,
-    './esm/Nullable/expect': null,
-    './esm/Nullable/inspect': null,
-    './esm/Nullable': null,
-    './esm/Nullable/map': null,
-    './esm/Nullable/mapAsync': null,
-    './esm/Nullable/mapOr': null,
-    './esm/Nullable/mapOrAsync': null,
-    './esm/Nullable/mapOrElse': null,
-    './esm/Nullable/mapOrElseAsync': null,
-    './esm/Nullable/namespace': null,
-    './esm/Nullable/okOr': null,
-    './esm/Nullable/okOrElse': null,
-    './esm/Nullable/or': null,
-    './esm/Nullable/orElse': null,
-    './esm/Nullable/orElseAsync': null,
-    './esm/Nullable/toPlainResult': null,
-    './esm/Nullable/toUndefinable': null,
-    './esm/Nullable/unwrap': null,
-    './esm/Nullable/unwrapOr': null,
-    './esm/Nullable/unwrapOrElse': null,
-    './esm/Nullable/unwrapOrElseAsync': null,
-    './esm/Nullable/xor': null,
-    './esm/PlainOption/Option': null,
-    './esm/PlainOption/and': null,
-    './esm/PlainOption/andThen': null,
-    './esm/PlainOption/andThenAsync': null,
-    './esm/PlainOption/asMut': null,
-    './esm/PlainOption/compat/v33': null,
-    './esm/PlainOption/drop': null,
-    './esm/PlainOption/equal': null,
-    './esm/PlainOption/expect': null,
-    './esm/PlainOption/filter': null,
-    './esm/PlainOption/flatten': null,
-    './esm/PlainOption/inspect': null,
-    './esm/PlainOption': null,
-    './esm/PlainOption/map': null,
-    './esm/PlainOption/mapAsync': null,
-    './esm/PlainOption/mapOr': null,
-    './esm/PlainOption/mapOrAsync': null,
-    './esm/PlainOption/mapOrElse': null,
-    './esm/PlainOption/mapOrElseAsync': null,
-    './esm/PlainOption/namespace': null,
-    './esm/PlainOption/okOr': null,
-    './esm/PlainOption/okOrElse': null,
-    './esm/PlainOption/or': null,
-    './esm/PlainOption/orElse': null,
-    './esm/PlainOption/orElseAsync': null,
-    './esm/PlainOption/transpose': null,
-    './esm/PlainOption/toNullable': null,
-    './esm/PlainOption/toUndefinable': null,
-    './esm/PlainOption/unwrap': null,
-    './esm/PlainOption/unwrapOr': null,
-    './esm/PlainOption/unwrapOrElse': null,
-    './esm/PlainOption/unwrapOrElseAsync': null,
-    './esm/PlainOption/xor': null,
-    './esm/PlainResult/Result': null,
-    './esm/PlainResult/and': null,
-    './esm/PlainResult/andThen': null,
-    './esm/PlainResult/andThenAsync': null,
-    './esm/PlainResult/asMut': null,
-    './esm/PlainResult/compat/v33': null,
-    './esm/PlainResult/drop': null,
-    './esm/PlainResult/equal': null,
-    './esm/PlainResult/expect': null,
-    './esm/PlainResult/flatten': null,
-    './esm/PlainResult/fromPromiseSettledResult': null,
-    './esm/PlainResult/inspect': null,
-    './esm/PlainResult': null,
-    './esm/PlainResult/map': null,
-    './esm/PlainResult/mapAsync': null,
-    './esm/PlainResult/mapErr': null,
-    './esm/PlainResult/mapErrAsync': null,
-    './esm/PlainResult/mapOr': null,
-    './esm/PlainResult/mapOrAsync': null,
-    './esm/PlainResult/mapOrElse': null,
-    './esm/PlainResult/mapOrElseAsync': null,
-    './esm/PlainResult/namespace': null,
-    './esm/PlainResult/or': null,
-    './esm/PlainResult/orElse': null,
-    './esm/PlainResult/orElseAsync': null,
-    './esm/PlainResult/transpose': null,
-    './esm/PlainResult/toNullable': null,
-    './esm/PlainResult/toOption': null,
-    './esm/PlainResult/toUndefinable': null,
-    './esm/PlainResult/tryCatch': null,
-    './esm/PlainResult/tryCatchAsync': null,
-    './esm/PlainResult/unwrap': null,
-    './esm/PlainResult/unwrapOr': null,
-    './esm/PlainResult/unwrapOrElse': null,
-    './esm/PlainResult/unwrapOrElseAsync': null,
-    './esm/PlainResult/unwrapOrThrowError': null,
-    './esm/Undefinable/ErrorMessage': null,
-    './esm/Undefinable/Undefinable': null,
-    './esm/Undefinable/and': null,
-    './esm/Undefinable/andThen': null,
-    './esm/Undefinable/andThenAsync': null,
-    './esm/Undefinable/compat/v33': null,
-    './esm/Undefinable/expect': null,
-    './esm/Undefinable/inspect': null,
-    './esm/Undefinable': null,
-    './esm/Undefinable/map': null,
-    './esm/Undefinable/mapAsync': null,
-    './esm/Undefinable/mapOr': null,
-    './esm/Undefinable/mapOrAsync': null,
-    './esm/Undefinable/mapOrElse': null,
-    './esm/Undefinable/mapOrElseAsync': null,
-    './esm/Undefinable/namespace': null,
-    './esm/Undefinable/okOr': null,
-    './esm/Undefinable/okOrElse': null,
-    './esm/Undefinable/or': null,
-    './esm/Undefinable/orElse': null,
-    './esm/Undefinable/orElseAsync': null,
-    './esm/Undefinable/toNullable': null,
-    './esm/Undefinable/toPlainResult': null,
-    './esm/Undefinable/unwrap': null,
-    './esm/Undefinable/unwrapOr': null,
-    './esm/Undefinable/unwrapOrElse': null,
-    './esm/Undefinable/unwrapOrElseAsync': null,
-    './esm/Undefinable/xor': null,
-    './lib/ClassicOption': null,
-    './lib/ClassicResult': null,
-    './lib/Maybe/ErrorMessage': null,
-    './lib/Maybe/Maybe': null,
-    './lib/Maybe/and': null,
-    './lib/Maybe/andThen': null,
-    './lib/Maybe/andThenAsync': null,
-    './lib/Maybe/compat/v33': null,
-    './lib/Maybe/expect': null,
-    './lib/Maybe/inspect': null,
-    './lib/Maybe': null,
-    './lib/Maybe/map': null,
-    './lib/Maybe/mapAsync': null,
-    './lib/Maybe/mapOr': null,
-    './lib/Maybe/mapOrAsync': null,
-    './lib/Maybe/mapOrElse': null,
-    './lib/Maybe/mapOrElseAsync': null,
-    './lib/Maybe/namespace': null,
-    './lib/Maybe/okOr': null,
-    './lib/Maybe/okOrElse': null,
-    './lib/Maybe/or': null,
-    './lib/Maybe/orElse': null,
-    './lib/Maybe/orElseAsync': null,
-    './lib/Maybe/toNullable': null,
-    './lib/Maybe/toPlainResult': null,
-    './lib/Maybe/toUndefinable': null,
-    './lib/Maybe/unwrap': null,
-    './lib/Maybe/unwrapOr': null,
-    './lib/Maybe/unwrapOrElse': null,
-    './lib/Maybe/unwrapOrElseAsync': null,
-    './lib/Maybe/xor': null,
-    './lib/Nullable/ErrorMessage': null,
-    './lib/Nullable/Nullable': null,
-    './lib/Nullable/and': null,
-    './lib/Nullable/andThen': null,
-    './lib/Nullable/andThenAsync': null,
-    './lib/Nullable/compat/v33': null,
-    './lib/Nullable/expect': null,
-    './lib/Nullable/inspect': null,
-    './lib/Nullable': null,
-    './lib/Nullable/map': null,
-    './lib/Nullable/mapAsync': null,
-    './lib/Nullable/mapOr': null,
-    './lib/Nullable/mapOrAsync': null,
-    './lib/Nullable/mapOrElse': null,
-    './lib/Nullable/mapOrElseAsync': null,
-    './lib/Nullable/namespace': null,
-    './lib/Nullable/okOr': null,
-    './lib/Nullable/okOrElse': null,
-    './lib/Nullable/or': null,
-    './lib/Nullable/orElse': null,
-    './lib/Nullable/orElseAsync': null,
-    './lib/Nullable/toPlainResult': null,
-    './lib/Nullable/toUndefinable': null,
-    './lib/Nullable/unwrap': null,
-    './lib/Nullable/unwrapOr': null,
-    './lib/Nullable/unwrapOrElse': null,
-    './lib/Nullable/unwrapOrElseAsync': null,
-    './lib/Nullable/xor': null,
-    './lib/PlainOption/Option': null,
-    './lib/PlainOption/and': null,
-    './lib/PlainOption/andThen': null,
-    './lib/PlainOption/andThenAsync': null,
-    './lib/PlainOption/asMut': null,
-    './lib/PlainOption/compat/v33': null,
-    './lib/PlainOption/drop': null,
-    './lib/PlainOption/expect': null,
-    './lib/PlainOption/equal': null,
-    './lib/PlainOption/flatten': null,
-    './lib/PlainOption/filter': null,
-    './lib/PlainOption/inspect': null,
-    './lib/PlainOption': null,
-    './lib/PlainOption/map': null,
-    './lib/PlainOption/mapAsync': null,
-    './lib/PlainOption/mapOr': null,
-    './lib/PlainOption/mapOrAsync': null,
-    './lib/PlainOption/mapOrElse': null,
-    './lib/PlainOption/mapOrElseAsync': null,
-    './lib/PlainOption/namespace': null,
-    './lib/PlainOption/okOr': null,
-    './lib/PlainOption/okOrElse': null,
-    './lib/PlainOption/or': null,
-    './lib/PlainOption/orElse': null,
-    './lib/PlainOption/orElseAsync': null,
-    './lib/PlainOption/transpose': null,
-    './lib/PlainOption/toNullable': null,
-    './lib/PlainOption/toUndefinable': null,
-    './lib/PlainOption/unwrap': null,
-    './lib/PlainOption/unwrapOr': null,
-    './lib/PlainOption/unwrapOrElse': null,
-    './lib/PlainOption/unwrapOrElseAsync': null,
-    './lib/PlainOption/xor': null,
-    './lib/PlainResult/Result': null,
-    './lib/PlainResult/and': null,
-    './lib/PlainResult/andThen': null,
-    './lib/PlainResult/andThenAsync': null,
-    './lib/PlainResult/asMut': null,
-    './lib/PlainResult/compat/v33': null,
-    './lib/PlainResult/drop': null,
-    './lib/PlainResult/equal': null,
-    './lib/PlainResult/expect': null,
-    './lib/PlainResult/flatten': null,
-    './lib/PlainResult/fromPromiseSettledResult': null,
-    './lib/PlainResult/inspect': null,
-    './lib/PlainResult': null,
-    './lib/PlainResult/map': null,
-    './lib/PlainResult/mapAsync': null,
-    './lib/PlainResult/mapErr': null,
-    './lib/PlainResult/mapErrAsync': null,
-    './lib/PlainResult/mapOr': null,
-    './lib/PlainResult/mapOrAsync': null,
-    './lib/PlainResult/mapOrElse': null,
-    './lib/PlainResult/mapOrElseAsync': null,
-    './lib/PlainResult/namespace': null,
-    './lib/PlainResult/or': null,
-    './lib/PlainResult/orElse': null,
-    './lib/PlainResult/orElseAsync': null,
-    './lib/PlainResult/transpose': null,
-    './lib/PlainResult/toNullable': null,
-    './lib/PlainResult/toOption': null,
-    './lib/PlainResult/toUndefinable': null,
-    './lib/PlainResult/tryCatch': null,
-    './lib/PlainResult/tryCatchAsync': null,
-    './lib/PlainResult/unwrap': null,
-    './lib/PlainResult/unwrapOr': null,
-    './lib/PlainResult/unwrapOrElse': null,
-    './lib/PlainResult/unwrapOrElseAsync': null,
-    './lib/PlainResult/unwrapOrThrowError': null,
-    './lib/Undefinable/ErrorMessage': null,
-    './lib/Undefinable/Undefinable': null,
-    './lib/Undefinable/and': null,
-    './lib/Undefinable/andThen': null,
-    './lib/Undefinable/andThenAsync': null,
-    './lib/Undefinable/expect': null,
-    './lib/Undefinable/compat/v33': null,
-    './lib/Undefinable/inspect': null,
-    './lib/Undefinable': null,
-    './lib/Undefinable/map': null,
-    './lib/Undefinable/mapAsync': null,
-    './lib/Undefinable/mapOr': null,
-    './lib/Undefinable/mapOrAsync': null,
-    './lib/Undefinable/mapOrElse': null,
-    './lib/Undefinable/mapOrElseAsync': null,
-    './lib/Undefinable/namespace': null,
-    './lib/Undefinable/okOr': null,
-    './lib/Undefinable/okOrElse': null,
-    './lib/Undefinable/or': null,
-    './lib/Undefinable/orElse': null,
-    './lib/Undefinable/orElseAsync': null,
-    './lib/Undefinable/toNullable': null,
-    './lib/Undefinable/toPlainResult': null,
-    './lib/Undefinable/unwrap': null,
-    './lib/Undefinable/unwrapOr': null,
-    './lib/Undefinable/unwrapOrElse': null,
-    './lib/Undefinable/unwrapOrElseAsync': null,
-    './lib/Undefinable/xor': null,
-    '.': null,
-    './Maybe': null,
-    './Maybe/and': null,
-    './Maybe/andThen': null,
-    './Maybe/andThenAsync': null,
-    './Maybe/compat/v33': null,
-    './Maybe/expect': null,
-    './Maybe/inspect': null,
-    './Maybe/map': null,
-    './Maybe/mapAsync': null,
-    './Maybe/Maybe': null,
-    './Maybe/mapOr': null,
-    './Maybe/mapOrAsync': null,
-    './Maybe/mapOrElse': null,
-    './Maybe/mapOrElseAsync': null,
-    './Maybe/namespace': null,
-    './Maybe/okOr': null,
-    './Maybe/okOrElse': null,
-    './Maybe/or': null,
-    './Maybe/orElse': null,
-    './Maybe/orElseAsync': null,
-    './Maybe/toNullable': null,
-    './Maybe/toPlainResult': null,
-    './Maybe/toUndefinable': null,
-    './Maybe/unwrap': null,
-    './Maybe/unwrapOr': null,
-    './Maybe/unwrapOrElse': null,
-    './Maybe/unwrapOrElseAsync': null,
-    './Maybe/xor': null,
-    './Nullable': null,
-    './Nullable/and': null,
-    './Nullable/andThen': null,
-    './Nullable/andThenAsync': null,
-    './Nullable/compat/v33': null,
-    './Nullable/expect': null,
-    './Nullable/inspect': null,
-    './Nullable/map': null,
-    './Nullable/mapAsync': null,
-    './Nullable/mapOr': null,
-    './Nullable/mapOrAsync': null,
-    './Nullable/mapOrElse': null,
-    './Nullable/mapOrElseAsync': null,
-    './Nullable/Nullable': null,
-    './Nullable/namespace': null,
-    './Nullable/okOr': null,
-    './Nullable/okOrElse': null,
-    './Nullable/or': null,
-    './Nullable/orElse': null,
-    './Nullable/orElseAsync': null,
-    './Nullable/toPlainResult': null,
-    './Nullable/toUndefinable': null,
-    './Nullable/unwrap': null,
-    './Nullable/unwrapOr': null,
-    './Nullable/unwrapOrElse': null,
-    './Nullable/unwrapOrElseAsync': null,
-    './Nullable/xor': null,
-    './PlainOption': null,
-    './PlainOption/and': null,
-    './PlainOption/andThen': null,
-    './PlainOption/andThenAsync': null,
-    './PlainOption/asMut': null,
-    './PlainOption/compat/v33': null,
-    './PlainOption/drop': null,
-    './PlainOption/equal': null,
-    './PlainOption/expect': null,
-    './PlainOption/filter': null,
-    './PlainOption/flatten': null,
-    './PlainOption/inspect': null,
-    './PlainOption/map': null,
-    './PlainOption/mapAsync': null,
-    './PlainOption/mapOr': null,
-    './PlainOption/mapOrAsync': null,
-    './PlainOption/mapOrElse': null,
-    './PlainOption/mapOrElseAsync': null,
-    './PlainOption/okOr': null,
-    './PlainOption/okOrElse': null,
-    './PlainOption/Option': null,
-    './PlainOption/namespace': null,
-    './PlainOption/or': null,
-    './PlainOption/orElse': null,
-    './PlainOption/orElseAsync': null,
-    './PlainOption/transpose': null,
-    './PlainOption/toNullable': null,
-    './PlainOption/toUndefinable': null,
-    './PlainOption/unwrap': null,
-    './PlainOption/unwrapOr': null,
-    './PlainOption/unwrapOrElse': null,
-    './PlainOption/unwrapOrElseAsync': null,
-    './PlainOption/xor': null,
-    './PlainResult': null,
-    './PlainResult/and': null,
-    './PlainResult/andThen': null,
-    './PlainResult/andThenAsync': null,
-    './PlainResult/asMut': null,
-    './PlainResult/compat/v33': null,
-    './PlainResult/drop': null,
-    './PlainResult/equal': null,
-    './PlainResult/expect': null,
-    './PlainResult/flatten': null,
-    './PlainResult/fromPromiseSettledResult': null,
-    './PlainResult/inspect': null,
-    './PlainResult/map': null,
-    './PlainResult/mapAsync': null,
-    './PlainResult/mapErr': null,
-    './PlainResult/mapErrAsync': null,
-    './PlainResult/mapOr': null,
-    './PlainResult/mapOrAsync': null,
-    './PlainResult/mapOrElse': null,
-    './PlainResult/mapOrElseAsync': null,
-    './PlainResult/namespace': null,
-    './PlainResult/or': null,
-    './PlainResult/orElse': null,
-    './PlainResult/orElseAsync': null,
-    './PlainResult/Result': null,
-    './PlainResult/toNullable': null,
-    './PlainResult/toOption': null,
-    './PlainResult/toUndefinable': null,
-    './PlainResult/tryCatch': null,
-    './PlainResult/tryCatchAsync': null,
-    './PlainResult/transpose': null,
-    './PlainResult/unwrap': null,
-    './PlainResult/unwrapOr': null,
-    './PlainResult/unwrapOrElse': null,
-    './PlainResult/unwrapOrElseAsync': null,
-    './PlainResult/unwrapOrThrowError': null,
-    './Undefinable': null,
-    './Undefinable/and': null,
-    './Undefinable/andThen': null,
-    './Undefinable/andThenAsync': null,
-    './Undefinable/expect': null,
-    './Undefinable/compat/v33': null,
-    './Undefinable/inspect': null,
-    './Undefinable/map': null,
-    './Undefinable/mapAsync': null,
-    './Undefinable/mapOr': null,
-    './Undefinable/mapOrAsync': null,
-    './Undefinable/mapOrElse': null,
-    './Undefinable/mapOrElseAsync': null,
-    './Undefinable/namespace': null,
-    './Undefinable/okOr': null,
-    './Undefinable/okOrElse': null,
-    './Undefinable/or': null,
-    './Undefinable/orElse': null,
-    './Undefinable/orElseAsync': null,
-    './Undefinable/toNullable': null,
-    './Undefinable/toPlainResult': null,
-    './Undefinable/Undefinable': null,
-    './Undefinable/unwrap': null,
-    './Undefinable/unwrapOr': null,
-    './Undefinable/unwrapOrElse': null,
-    './Undefinable/unwrapOrElseAsync': null,
-    './Undefinable/xor': null,
-});
